+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Show date ranges upon backtesting
 - Show portfolio at the end of backtesting
+- Automatically export candles and graph data to CSV
 
 ---